Use Xen console_io hypercalls for console output
When running as dom0less guest, the xen_hvc interface is not available
and an emulated PL011 serial is provided.

Instead of adding the PL011 driver to Zephyr, use the console_io
hypercalls for output which are available for all guests to use in Xen
DEBUG builds.

The console_io interface is already supported in Zephyr. We only need
to:
- configure Zephyr to be built to run as dom0, even if we are going to
  run it as domU
- rename xen_hvc to xen_consoleio_hvc in xenvm.dts, which triggers the
  usage of the console_io interface

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>