Add a timer latency example to hello_world
Add a Xilinx TTC timer device to the xenvm machine, both to xenvm.dts
and to the memory mappings listed in mmu_regions.c. The Xilinx TTC
device will be directly assigned to Zephyr.

Add a small example to hello_world that uses the TTC timer and prints a
timestamp every few seconds.

Allow the xenvm machine to use the TTC timer instead of the ARM Arch
Timer in the kconfig.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>